+++ apm-agent-api/src/main/java/co/elastic/apm/api/SpanImpl.java
@@ -56,15 +56,22 @@ class SpanImpl extends AbstractSpanImpl {
 
     @Nonnull
     @Override
-    public Span addTag(String key, Number value) {
-        doAddNumberTag(key, value);
+    public Span addLabel(String key, String value) {
+        doAddTag(key, value);
         return this;
     }
 
     @Nonnull
     @Override
-    public Span addTag(String key, boolean value) {
-        doAddBooleanTag(key, value);
+    public Span addLabel(String key, Number value) {
+        doAddNumberLabel(key, value);
+        return this;
+    }
+
+    @Nonnull
+    @Override
+    public Span addLabel(String key, boolean value) {
+        doAddBooleanLabel(key, value);
         return this;
     }