+++ apm-agent-plugins/apm-api-plugin/src/main/java/co/elastic/apm/agent/plugin/api/AbstractSpanInstrumentation.java
@@ -173,45 +173,45 @@ public class AbstractSpanInstrumentation extends ApiInstrumentation {
         }
     }
 
-    public static class AddTagInstrumentation extends AbstractSpanInstrumentation {
-        public AddTagInstrumentation() {
+    public static class AddStringLabelInstrumentation extends AbstractSpanInstrumentation {
+        public AddStringLabelInstrumentation() {
             super(named("doAddTag"));
         }
 
         @VisibleForAdvice
         @Advice.OnMethodEnter(suppress = Throwable.class)
-        public static void addTag(@Advice.FieldValue(value = "span", typing = Assigner.Typing.DYNAMIC) TraceContextHolder<?> context,
-                                  @Advice.Argument(0) String key, @Nullable @Advice.Argument(1) String value) {
+        public static void addLabel(@Advice.FieldValue(value = "span", typing = Assigner.Typing.DYNAMIC) TraceContextHolder<?> context,
+                                    @Advice.Argument(0) String key, @Nullable @Advice.Argument(1) String value) {
             if (value != null && context instanceof AbstractSpan) {
                 ((AbstractSpan) context).addTag(key, value);
             }
         }
     }
 
-    public static class AddNumberTagInstrumentation extends AbstractSpanInstrumentation {
-        public AddNumberTagInstrumentation() {
-            super(named("doAddNumberTag"));
+    public static class AddNumberLabelInstrumentation extends AbstractSpanInstrumentation {
+        public AddNumberLabelInstrumentation() {
+            super(named("doAddNumberLabel"));
         }
 
         @VisibleForAdvice
         @Advice.OnMethodEnter(suppress = Throwable.class)
-        public static void addTag(@Advice.FieldValue(value = "span", typing = Assigner.Typing.DYNAMIC) TraceContextHolder<?> context,
-                                  @Advice.Argument(0) String key, @Nullable @Advice.Argument(1) Number value) {
+        public static void addLabel(@Advice.FieldValue(value = "span", typing = Assigner.Typing.DYNAMIC) TraceContextHolder<?> context,
+                                    @Advice.Argument(0) String key, @Nullable @Advice.Argument(1) Number value) {
             if (value != null && context instanceof AbstractSpan) {
                 ((AbstractSpan) context).addTag(key, value);
             }
         }
     }
 
-    public static class AddBooleanTagInstrumentation extends AbstractSpanInstrumentation {
-        public AddBooleanTagInstrumentation() {
-            super(named("doAddBooleanTag"));
+    public static class AddBooleanLabelInstrumentation extends AbstractSpanInstrumentation {
+        public AddBooleanLabelInstrumentation() {
+            super(named("doAddBooleanLabel"));
         }
 
         @VisibleForAdvice
         @Advice.OnMethodEnter(suppress = Throwable.class)
-        public static void addTag(@Advice.FieldValue(value = "span", typing = Assigner.Typing.DYNAMIC) TraceContextHolder<?> context,
-                                  @Advice.Argument(0) String key, @Nullable @Advice.Argument(1) Boolean value) {
+        public static void addLabel(@Advice.FieldValue(value = "span", typing = Assigner.Typing.DYNAMIC) TraceContextHolder<?> context,
+                                    @Advice.Argument(0) String key, @Nullable @Advice.Argument(1) Boolean value) {
             if (value != null && context instanceof AbstractSpan) {
                 ((AbstractSpan) context).addTag(key, value);
             }