+++ apm-agent-plugins/apm-api-plugin/src/test/java/co/elastic/apm/api/AnnotationApiTest.java
@@ -34,12 +34,12 @@ class AnnotationApiTest extends AbstractInstrumentationTest {
 
         assertThat(reporter.getTransactions()).hasSize(1);
         assertThat(reporter.getFirstTransaction().getName().toString()).isEqualTo("transaction");
-        assertThat(reporter.getFirstTransaction().getContext().getTag("foo")).isEqualTo("bar");
+        assertThat(reporter.getFirstTransaction().getContext().getLabel("foo")).isEqualTo("bar");
 
         assertThat(reporter.getSpans()).hasSize(2);
 
         assertThat(reporter.getSpans().get(0).getName().toString()).isEqualTo("AnnotationTestClass#nestedSpan");
-        assertThat(reporter.getSpans().get(0).getContext().getTag("foo")).isEqualTo("bar");
+        assertThat(reporter.getSpans().get(0).getContext().getLabel("foo")).isEqualTo("bar");
         assertThat(reporter.getSpans().get(0).isChildOf(reporter.getSpans().get(1))).isTrue();
 
         assertThat(reporter.getSpans().get(1).getName().toString()).isEqualTo("AnnotationTestClass#span");
@@ -97,7 +97,7 @@ class AnnotationApiTest extends AbstractInstrumentationTest {
 
         @CaptureTransaction("transaction")
         void transaction() {
-            ElasticApm.currentTransaction().addTag("foo", "bar");
+            ElasticApm.currentTransaction().addLabel("foo", "bar");
             span();
         }
 
@@ -108,7 +108,7 @@ class AnnotationApiTest extends AbstractInstrumentationTest {
 
         @CaptureSpan
         void nestedSpan() {
-            ElasticApm.currentSpan().addTag("foo", "bar");
+            ElasticApm.currentSpan().addLabel("foo", "bar");
         }
 
         @CaptureTransaction(value = "transactionWithType", type = "job")