+++ apm-agent-plugins/apm-api-plugin/src/test/java/co/elastic/apm/api/ElasticApmApiInstrumentationTest.java
@@ -155,32 +155,36 @@ class ElasticApmApiInstrumentationTest extends AbstractInstrumentationTest {
     }
 
     @Test
-    void testAddTag() {
+    void testAddLabel() {
         Transaction transaction = ElasticApm.startTransaction();
         transaction.setName("foo");
         transaction.setType("bar");
-        transaction.addTag("foo", "bar");
-        transaction.addTag("number", 1);
-        transaction.addTag("boolean", true);
-        transaction.addTag("null", (String) null);
+        transaction.addLabel("foo1", "bar1");
+        transaction.addLabel("foo", "bar");
+        transaction.addLabel("number", 1);
+        transaction.addLabel("boolean", true);
+        transaction.addLabel("null", (String) null);
         Span span = transaction.startSpan("bar", null, null);
         span.setName("foo");
-        span.addTag("bar", "baz");
-        span.addTag("number", 1);
-        span.addTag("boolean", true);
-        span.addTag("null", (String) null);
+        span.addLabel("bar1", "baz1");
+        span.addLabel("bar", "baz");
+        span.addLabel("number", 1);
+        span.addLabel("boolean", true);
+        span.addLabel("null", (String) null);
         span.end();
         transaction.end();
         assertThat(reporter.getTransactions()).hasSize(1);
         assertThat(reporter.getSpans()).hasSize(1);
-        assertThat(reporter.getFirstTransaction().getContext().getTag("foo")).isEqualTo("bar");
-        assertThat(reporter.getFirstTransaction().getContext().getTag("number")).isEqualTo(1);
-        assertThat(reporter.getFirstTransaction().getContext().getTag("boolean")).isEqualTo(true);
-        assertThat(reporter.getFirstTransaction().getContext().getTag("null")).isNull();
-        assertThat(reporter.getFirstSpan().getContext().getTag("bar")).isEqualTo("baz");
-        assertThat(reporter.getFirstSpan().getContext().getTag("number")).isEqualTo(1);
-        assertThat(reporter.getFirstSpan().getContext().getTag("boolean")).isEqualTo(true);
-        assertThat(reporter.getFirstSpan().getContext().getTag("null")).isNull();
+        assertThat(reporter.getFirstTransaction().getContext().getLabel("foo1")).isEqualTo("bar1");
+        assertThat(reporter.getFirstTransaction().getContext().getLabel("foo")).isEqualTo("bar");
+        assertThat(reporter.getFirstTransaction().getContext().getLabel("number")).isEqualTo(1);
+        assertThat(reporter.getFirstTransaction().getContext().getLabel("boolean")).isEqualTo(true);
+        assertThat(reporter.getFirstTransaction().getContext().getLabel("null")).isNull();
+        assertThat(reporter.getFirstSpan().getContext().getLabel("bar1")).isEqualTo("baz1");
+        assertThat(reporter.getFirstSpan().getContext().getLabel("bar")).isEqualTo("baz");
+        assertThat(reporter.getFirstSpan().getContext().getLabel("number")).isEqualTo(1);
+        assertThat(reporter.getFirstSpan().getContext().getLabel("boolean")).isEqualTo(true);
+        assertThat(reporter.getFirstSpan().getContext().getLabel("null")).isNull();
     }
 
     @Test